+++ azurerm/resources.go
@@ -1,0 +1,135 @@
+package azurerm
+
+import (
+	"context"
+
+	"github.com/pkg/errors"
+
+	"github.com/cycloidio/terracognita/provider"
+	"github.com/cycloidio/terracognita/tag"
+)
+
+// ResourceType is the type used to define all the Resources
+// from the Provider
+type ResourceType int
+
+//go:generate enumer -type ResourceType -addprefix azurerm_ -transform snake -linecomment
+const (
+	ResourceGroup ResourceType = iota
+	VirtualMachine
+	VirtualNetwork
+	Subnet
+	NetworkInterface
+	NetworkSecurityGroup
+	VirtualMachineScaleSet
+)
+
+type rtFn func(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error)
+
+var (
+	resources = map[ResourceType]rtFn{
+		ResourceGroup:          resourceGroup,
+		VirtualMachine:         virtualMachines,
+		VirtualNetwork:         cacheVirtualNetworks,
+		Subnet:                 subnets,
+		NetworkInterface:       networkInterfaces,
+		NetworkSecurityGroup:   networkSecurityGroups,
+		VirtualMachineScaleSet: virtualMachineScaleSets,
+	}
+)
+
+func resourceGroup(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	resourceGroup := a.azurer.GetResourceGroup()
+	r := provider.NewResource(*resourceGroup.ID, resourceType, a)
+	resources := []provider.Resource{r}
+	return resources, nil
+}
+
+func virtualMachines(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	virtualMachines, err := a.azurer.ListVirtualMachines(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list virtual machines from reader")
+	}
+	resources := make([]provider.Resource, 0, len(virtualMachines))
+	for _, virtualMachine := range virtualMachines {
+		r := provider.NewResource(*virtualMachine.ID, resourceType, a)
+		resources = append(resources, r)
+	}
+	return resources, nil
+}
+
+func virtualNetworks(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	virtualNetworks, err := a.azurer.ListVirtualNetworks(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list virtual networks from reader")
+	}
+	resources := make([]provider.Resource, 0, len(virtualNetworks))
+	for _, virtualNetwork := range virtualNetworks {
+		r := provider.NewResource(*virtualNetwork.ID, resourceType, a)
+		// we set the name prior of reading it from the state
+		// as it is required to able to List resources depending on this one
+		if err := r.Data().Set("name", *virtualNetwork.Name); err != nil {
+			return nil, errors.Wrapf(err, "unable to set name data on the provider.Resource for the virtual network '%s'", *virtualNetwork.Name)
+		}
+		resources = append(resources, r)
+	}
+	return resources, nil
+}
+
+func subnets(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	virtualNetworkNames, err := getVirtualNetworkNames(ctx, a, resourceType, tags)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list virtual networks from cache")
+	}
+	resources := make([]provider.Resource, 0)
+	for _, virtualNetworkName := range virtualNetworkNames {
+		subnets, err := a.azurer.ListSubnets(ctx, virtualNetworkName)
+		if err != nil {
+			return nil, errors.Wrap(err, "unable to list subnets from reader")
+		}
+		for _, subnet := range subnets {
+			r := provider.NewResource(*subnet.ID, resourceType, a)
+			resources = append(resources, r)
+		}
+	}
+	return resources, nil
+}
+
+func networkInterfaces(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	networkInterfaces, err := a.azurer.ListInterfaces(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list network interfaces from reader")
+	}
+	resources := make([]provider.Resource, 0, len(networkInterfaces))
+	for _, networkInterface := range networkInterfaces {
+		r := provider.NewResource(*networkInterface.ID, resourceType, a)
+		resources = append(resources, r)
+	}
+	return resources, nil
+}
+
+func networkSecurityGroups(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	securityGroups, err := a.azurer.ListSecurityGroups(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list network security groups from reader")
+	}
+	resources := make([]provider.Resource, 0, len(securityGroups))
+	for _, securityGroup := range securityGroups {
+		r := provider.NewResource(*securityGroup.ID, resourceType, a)
+		resources = append(resources, r)
+	}
+	return resources, nil
+}
+
+func virtualMachineScaleSets(ctx context.Context, a *azurerm, resourceType string, tags []tag.Tag) ([]provider.Resource, error) {
+	virtualMachineScaleSets, err := a.azurer.ListVirtualMachineScaleSets(ctx)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to list virtual machines scale sets from reader")
+	}
+	resources := make([]provider.Resource, 0, len(virtualMachineScaleSets))
+	for _, virtualMachineScaleSet := range virtualMachineScaleSets {
+		r := provider.NewResource(*virtualMachineScaleSet.ID, resourceType, a)
+		resources = append(resources, r)
+	}
+	return resources, nil
+}